Updates to project architecture & add shape drawer dependency
Entities will be passed the world object so they are able to update
their own position based on results from the jbump world simulation.
This saves having a bunch of unruly code bloating the level/world class.

Render methods now live the Entity class to save having many defs in
each extending class.

Added shape drawer, this allows shapes to be drawn using the sprite
batch so we don't need to flush the batch to use a shapeRenderer for
bounding box outlines.